+++ packages/headless/src/features/triggers/triggers-state.ts
@@ -8,9 +8,15 @@ export interface TriggerState {
    * The new query to perform a search with after receiving a query trigger.
    */
   query: string;
+
+  /**
+   * The notification to display to the user after receiving a notification trigger.
+   */
+  notification: string;
 }
 
 export const getTriggerInitialState: () => TriggerState = () => ({
   redirectTo: '',
   query: '',
+  notification: '',
 });